Initial Code Structure
Wrote server and client code. Able to start a server and connect the client, register a handle, list the uploaded files, and get a file from the uploads. However, uploading files through the interface is still not working properly. Storing a text file only creates a new file under 'uploads' with no contents.